Util.c e util.h
