+++ Makefile
@@ -1,11 +1,14 @@
-all: funcoes_fornecidas.o reader.o regCabecalho.o indiceSimples.o programaTrab.o
-	gcc funcoes_fornecidas.o reader.o regCabecalho.o indiceSimples.o programaTrab.o -Wall -std=c99 -o programaTrab
+all: funcoes_fornecidas.o reader.o regCabecalho.o indiceSimples.o util.o programaTrab.o
+	gcc funcoes_fornecidas.o util.o reader.o regCabecalho.o indiceSimples.o programaTrab.o -Wall -std=c99 -o programaTrab
 
 run: all
 	./programaTrab > test.txt
 
 programaTrab.o: programaTrab.c
 	gcc -c programaTrab.c -o programaTrab.o
+
+util.o: util.c
+	gcc -c util.c -o util.o
 
 reader.o: reader.c reader.h
 	gcc -c reader.c -o reader.o
